Curl Scripts done for the Events schema, few implementations still needed for RSVP.
Changes were made on the events folder for curl scripts. The create, destroy and index needed to be re-done because it was not working properly and the updated sh had the wrong field name. Now everything is working fine, but there is still needs to implement the RSVP for the schema.  File edited by Leni Lunderman.